Complete tasks of sprint 7
