+++ test/managerTest/FileBackedTaskManagerTest.java
@@ -1,6 +1,7 @@
 package managerTest;
 
 import file.FileBackedTaskManager;
+import file.InMemoryTaskManager;
 import file.ManagerSaveException;
 import org.junit.jupiter.api.Test;
 import task.Epic;
@@ -13,11 +14,17 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Writer;
 import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class FileBackedTaskManagerTest {
+class FileBackedTaskManagerTest extends TaskManagerTest<FileBackedTaskManager> {
     FileBackedTaskManager file = new FileBackedTaskManager();
+
+    public FileBackedTaskManagerTest() {
+        super(new FileBackedTaskManager());
+    }
 
     @Test
     void saveTaskInFileAndReadFromFile() {
@@ -43,115 +50,9 @@ class FileBackedTaskManagerTest {
     }
 
     @Test
-    void addTask() {
-        Task task = new Task(file.getNewID(), "1", "1");
-        file.addTask(task);
-        assertEquals(task, file.getTaskByld(task.getID()), "Задача не добавлена");
-    }
-
-    @Test
-    void addEpic() {
-        Epic epic = new Epic(file.getNewID(), "1", "1");
-        file.addEpic(epic);
-        assertEquals(epic, file.getEpicByld(epic.getID()), "Задача не добавлена");
-    }
-
-    @Test
-    void addSubtask() {
-        file.addEpic(new Epic(file.getNewID(), "1", "1"));
-        Subtask subtask = new Subtask(file.getNewID(), "111", "111", 1);
-        file.addSubtask(subtask);
-        assertEquals(subtask, file.getSubtaskByld(subtask.getID()), "Задача не добавлена");
-    }
-
-    @Test
-    void clearTasks() {
-        Task task = new Task(file.getNewID(), "1", "1");
-        file.addTask(task);
-        file.clearTasks();
-        assertTrue(file.getTasks().isEmpty(), "Список задач не очистился");
-    }
-
-    @Test
-    void clearEpics() {
-        Epic epic = new Epic(file.getNewID(), "1", "1");
-        file.addEpic(epic);
-        Subtask subtask = new Subtask(file.getNewID(), "222", "222", 1);
-        file.addSubtask(subtask);
-        file.clearEpics();
-        assertTrue(file.getEpics().isEmpty() && file.getSubtasks().isEmpty(),
-                "Списки суперзадач и подзадач не очищены");
-    }
-
-    @Test
-    void clearSubtasks() {
-        Epic epic = new Epic(file.getNewID(), "1", "1");
-        file.addEpic(epic);
-        Subtask subtask = new Subtask(file.getNewID(), "111", "111", 1);
-        file.addSubtask(subtask);
-        file.clearSubtasks();
-        assertTrue(file.getSubtasks().isEmpty() && epic.getSubtaskList().isEmpty(),
-                "Список подзадач не очищен");
-    }
-
-    @Test
-    void removeTaskOnID() {
-        Task task = new Task(file.getNewID(), "1", "1");
-        file.addTask(task);
-        file.getTaskByld(1);
-        file.removeTaskOnID(task.getID());
-        assertFalse(file.getTasks().containsKey(task.getID()),
-                "Задача не была удалена из списка");
-    }
-
-    @Test
-    void removeEpicOnID() {
-        Epic epic = new Epic(file.getNewID(), "1", "1");
-        file.addEpic(epic);
-        file.getEpicByld(1);
-        file.removeEpicOnID(epic.getID());
-        assertFalse(file.getEpics().containsKey(epic.getID()), "Суперзадача не была удалена");
-    }
-
-    @Test
-    void removeSubtaskOnID() {
-        Epic epic = new Epic(file.getNewID(), "1", "1");
-        file.addEpic(epic);
-        Subtask subtask = new Subtask(file.getNewID(), "111", "111", 1);
-        file.addSubtask(subtask);
-        file.getEpicByld(1);
-        file.getSubtaskByld(2);
-        file.removeSubtaskOnID(subtask.getID());
-        assertFalse(file.getSubtasks().containsKey(subtask.getID())
-                        && epic.getSubtaskList().contains(subtask),
-                "Подзадача не была удалена из списка подзадач и списка подзадач суперзадачи");
-    }
-
-    @Test
-    void updateTask() {
-        Task task = new Task(file.getNewID(), "1", "1");
-        file.addTask(task);
-        file.updateTask(new Task(1, "2", "2"));
-        assertNotEquals(task, file.getTaskByld(task.getID()), "Задача не обновилась");
-    }
-
-    @Test
-    void updateEpic() {
-        Epic epic = new Epic(file.getNewID(), "1", "1");
-        file.addEpic(epic);
-        file.updateEpic(new Epic(1, "2", "2"));
-        assertNotEquals(epic, file.getEpicByld(epic.getID()), "Суперзадача не обновилась");
-    }
-
-    @Test
-    void updateSubtask() {
-        Epic epic = new Epic(file.getNewID(), "1", "1");
-        file.addEpic(epic);
-        Subtask subtask = new Subtask(file.getNewID(), "111", "111", 1);
-        file.addSubtask(subtask);
-        file.updateSubtask(new Subtask(2, "2", "2", 1,
-                TaskStatus.DONE));
-        assertFalse(subtask.equals(file.getSubtaskByld(subtask.getID()))
-                && subtask.equals(epic.getSubtaskList().getFirst()), "Подзадача не обновилась");
+    void CorrectlyCatchExceptions() {
+        assertThrows(NullPointerException.class, () -> {
+            manager.addTask(null);
+        }, "Ожидается NullPointerException при ошибке записи в файл");
     }
 }
